requirements.txt
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.9.2
Pillow==9.4.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
translate==3.6.1
urllib3==1.26.14
